refactor: move types to .d.ts files in preparation of ts-to-jsdoc
