+++ packages/vite-plugin-svelte/src/preprocess.d.ts
@@ -1,0 +1,6 @@
+export type CssTransform = (
+	// eslint-disable-next-line no-unused-vars
+	code: string,
+	// eslint-disable-next-line no-unused-vars
+	filename: string
+) => Promise<{ code: string; map?: any; deps?: Set<string> }>;
