+++ packages/vite-plugin-svelte/src/utils/compile.d.ts
@@ -1,0 +1,44 @@
+// eslint-disable-next-line node/no-missing-import
+import type { Processed } from 'svelte/types/compiler/preprocess';
+import { _createCompileSvelte } from './compile';
+
+export type CompileSvelte = ReturnType<typeof _createCompileSvelte>;
+
+export interface Code {
+	code: string;
+	map?: any;
+	dependencies?: any[];
+}
+
+export interface Compiled {
+	js: Code;
+	css: Code;
+	ast: any; // TODO type
+	warnings: any[]; // TODO type
+	vars: {
+		name: string;
+		export_name: string;
+		injected: boolean;
+		module: boolean;
+		mutated: boolean;
+		reassigned: boolean;
+		referenced: boolean;
+		writable: boolean;
+		referenced_from_script: boolean;
+	}[];
+	stats: {
+		timings: {
+			total: number;
+		};
+	};
+}
+
+export interface CompileData {
+	filename: string;
+	normalizedFilename: string;
+	lang: string;
+	compiled: Compiled;
+	ssr: boolean | undefined;
+	dependencies: string[];
+	preprocessed: Processed;
+}
