+++ packages/vite-plugin-svelte/src/utils/compile.ts
@@ -1,21 +1,19 @@
-import { CompileOptions, ResolvedOptions } from './options';
+import { CompileOptions, ResolvedOptions } from './options.d';
 import { compile, preprocess, walk } from 'svelte/compiler';
 // @ts-ignore
 import { createMakeHot } from 'svelte-hmr';
 import { SvelteRequest } from './id';
 import { safeBase64Hash } from './hash';
 import { log } from './log';
-import { StatCollection } from './vite-plugin-svelte-stats';
 //eslint-disable-next-line node/no-missing-import
-import type { Processed } from 'svelte/types/compiler/preprocess';
 import { createInjectScopeEverythingRulePreprocessorGroup } from './preprocess';
 import { mapToRelative } from './sourcemaps';
+import type { CompileData } from './compile.d';
+import type { StatCollection } from './vite-plugin-svelte-stats.d';
 
 const scriptLangRE = /<script [^>]*lang=["']?([^"' >]+)["']?[^>]*>/;
 
-export type CompileSvelte = ReturnType<typeof _createCompileSvelte>;
-
-const _createCompileSvelte = (makeHot: Function) => {
+export const _createCompileSvelte = (makeHot: Function) => {
 	let stats: StatCollection | undefined;
 	const devStylePreprocessor = createInjectScopeEverythingRulePreprocessorGroup();
 	return async function compileSvelte(
@@ -190,43 +188,4 @@ function buildMakeHot(options: ResolvedOptions) {
 export function createCompileSvelte(options: ResolvedOptions) {
 	const makeHot = buildMakeHot(options);
 	return _createCompileSvelte(makeHot);
-}
-
-export interface Code {
-	code: string;
-	map?: any;
-	dependencies?: any[];
-}
-
-export interface Compiled {
-	js: Code;
-	css: Code;
-	ast: any; // TODO type
-	warnings: any[]; // TODO type
-	vars: {
-		name: string;
-		export_name: string;
-		injected: boolean;
-		module: boolean;
-		mutated: boolean;
-		reassigned: boolean;
-		referenced: boolean;
-		writable: boolean;
-		referenced_from_script: boolean;
-	}[];
-	stats: {
-		timings: {
-			total: number;
-		};
-	};
-}
-
-export interface CompileData {
-	filename: string;
-	normalizedFilename: string;
-	lang: string;
-	compiled: Compiled;
-	ssr: boolean | undefined;
-	dependencies: string[];
-	preprocessed: Processed;
 }
