+++ packages/vite-plugin-svelte/src/utils/dependencies.d.ts
@@ -1,0 +1,4 @@
+export interface DependencyData {
+	dir: string;
+	pkg: Record<string, any>;
+}
