+++ packages/vite-plugin-svelte/src/utils/sourcemaps.d.ts
@@ -1,0 +1,5 @@
+export interface SourceMapFileRefs {
+	file?: string;
+	sources?: string[];
+	sourceRoot?: string;
+}
