+++ packages/vite-plugin-svelte/src/utils/vite-plugin-svelte-cache.d.ts
@@ -1,0 +1,6 @@
+export interface PackageInfo {
+	name: string;
+	version: string;
+	svelte?: string;
+	path: string;
+}
